Wallet Points Fixes

+++ app/src/main/java/com/lionscare/app/data/repositories/wallet/response/GetBalanceResponse.kt
@@ -14,7 +14,7 @@ data class GetBalanceResponse(
 @Keep
 data class BalanceData(
     val id: String? = null,
-    val raw_value: Int? = 0,
+    val raw_value: Double? = 0.0,
     val value: String? = null,
     val group_name: String? = null
 )
